engine: disable OpenSSL::Engine with OpenSSL 3.0
OpenSSL::Engine will not be defined when compiled with OpenSSL 3.0.

The ENGINE API is deprecated in OpenSSL 3.0, in favor of the new
"Provider" concept.

We will add ways to interact with providers from Ruby programs, but
since the Provider concept is completely different from the ENGINE API,
it is not possible to implement a compatibility layer with the current
OpenSSL::Engine.