digest: use EVP_MD_CTX_get0_md() instead of EVP_MD_CTX_md() if exists
The function was renamed in OpenSSL 3.0 due to the change of the
lifetime of an EVP_MD object. It is no longer always statically
allocated and is now reference-counted - when the associated EVP_MD_CTX
is free'd, the EVP_MD can also become inaccessible.

Currently Ruby/OpenSSL only handles builtin algorithms, so no special
handling is needed except for adapting to the rename.

Note that OpenSSL::Digest#reset is changed to simply use NULL in place
of an EVP_MD_CTX_md() call. It was not necessary as they are not
changing the type and just clearing state.